t/t0021: convert the rot13-filter.pl script to C
This script is currently used by three test files: t0021-conversion.sh,
t2080-parallel-checkout-basics.sh, and
t2082-parallel-checkout-attributes.sh. To avoid the need for the PERL
dependency at these tests, let's convert the script to a C test-tool
command. Note, however, that we still use the script as a wrapper at
this commit, in order to minimize the amount of changes it introduces
and help reviewers. At the next commit we will properly remove the
script and adjust the affected tests to use test-tool.

Furthermore, note that there is a small adjustment at test
t0021-conversion.sh because it depended on a specific error message
given by perl's die routine.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>
Signed-off-by: Junio C Hamano <gitster@pobox.com>